+++ src/app/api/trpc/[trpc]/route.ts
@@ -1,25 +1,13 @@
 import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
-import { type NextRequest } from "next/server";
-
-import { appRouter } from "server/api/root";
-import { createTRPCContext } from "server/api/trpc";
-
-/**
- * This wraps the `createTRPCContext` helper and provides the required context for the tRPC API when
- * handling a HTTP request (e.g. when you make requests from Client Components).
- */
-const createContext = async (req: NextRequest) => {
-    return createTRPCContext({
-        headers: req.headers,
-    });
-};
+import { NextRequest } from "next/server";
+import { appRouter } from "server/routers/_app";
 
 const handler = (req: NextRequest) =>
     fetchRequestHandler({
         endpoint: "/api/trpc",
         req,
         router: appRouter,
-        createContext: () => createContext(req),
+        createContext: () => ({}),
         onError:
             process.env.NODE_ENV === "development"
                 ? ({ path, error }) => {
