+++ src/server/routers/group.router.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { procedure, router } from "../trpc";
+import { procedure, router } from "server/api/trpc";
 
 import { Group } from "models";
 import { TRPCError } from "@trpc/server";
