+++ src/server/routers/webhook.router.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { procedure, router } from "../trpc";
+import { procedure, router } from "server/api/trpc";
 import { TRPCError } from "@trpc/server";
 
 const webhookRouter = router({
